feat(shared.lua): `vim.traverse`
Might be jumping the gun on this (not very critical), but it was
low-cost to just post it.

It replaces the lost recursive functionality from #25118 and mainly it
would just exist to nudge people away from coming up with something
worse.

Note: Followed advice from
https://github.com/neovim/neovim/pull/25118#issuecomment-1717298984 to
put the callback arg last, although this means the optional arg isn't
last.